+++ Reddit/Easy/closest_string/src/lib.rs
@@ -6,3 +6,10 @@ pub fn hamming_distance(one: &str, two: &str) -> usize {
 
     one.chars().zip(two.chars()).filter(|&(a, b)| a != b).count()
 }
+
+pub fn closest_string(sequences: Vec<String>) -> String {
+    println!("{:?}", sequences);
+
+    unimplemented!();
+}
+